updated html builder, html tool structure
