+++ backpipe/tools/__init__.py
@@ -1,1 +1,2 @@
 import backpipe.tools.json as json
+import backpipe.tools.html as html
